split the main in a separate file and add some final touches in cli
